aviator_bot_backend/requirements/common.txt
Django==4.1.6
djangorestframework==3.14.0
psycopg2-binary>=2.8,<2.9